+++ src/main/java/org/dinosaur/foodbowl/domain/review/dto/request/ReviewCreateRequest.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.review.application.dto.request;
+package org.dinosaur.foodbowl.domain.review.dto.request;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
