+++ src/test/java/org/dinosaur/foodbowl/domain/review/application/ReviewServiceTest.java
@@ -9,9 +9,9 @@ import java.util.Collections;
 import java.util.List;
 import org.dinosaur.foodbowl.domain.member.domain.Member;
 import org.dinosaur.foodbowl.domain.photo.domain.Photo;
-import org.dinosaur.foodbowl.domain.review.application.dto.request.ReviewCreateRequest;
-import org.dinosaur.foodbowl.domain.review.application.dto.request.ReviewUpdateRequest;
 import org.dinosaur.foodbowl.domain.review.domain.Review;
+import org.dinosaur.foodbowl.domain.review.dto.request.ReviewCreateRequest;
+import org.dinosaur.foodbowl.domain.review.dto.request.ReviewUpdateRequest;
 import org.dinosaur.foodbowl.domain.review.persistence.ReviewRepository;
 import org.dinosaur.foodbowl.global.exception.BadRequestException;
 import org.dinosaur.foodbowl.global.exception.NotFoundException;
